% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixcovid19.r
\name{calculate_lag_data}
\alias{calculate_lag_data}
\title{Calculate Lag Dataframe}
\usage{
calculate_lag_data(data, season_target, seasons, current_dates, threshold)
}
\arguments{
\item{data}{Dataframe. See \code{details}}

\item{season_target}{Character. Target Season.}

\item{seasons}{Character Vector. A vector with the previous Seasons.}

\item{current_dates}{Numeric Vector. Dates in which you want to launch the simulation. Preferable more than one.}

\item{threshold}{Numeric. How high want the accurate alignment. Less value more precision.}
}
\value{
Returns a new dataframe with the optimal week lag and is composed by:

\describe{
  \item{season}{\code{character} class. A tibble with the season detail}
  \item{week_lag}{\code{numeric} class. A tibble with the week lag}
}
}
\description{
This function is the core of the Covid19 project. It works as showed

  1. It prepares the data in weekly date and calculate the cumulative of the orders quantity in this function \code{\link{prepare_data_weekly}}.

  2. It slices the seasons not season_target from the the first observation to the second observation after the last observation of the season_target season in this function \code{\link{slice_data}}.

  3. It uses the `dtw` package to make a dynamic time warping between the season_target season and the others. Then calculate the optimal lag between them in this function \code{\link{dtw_align_curves}}.

  4. It stores the lag data in Dataframe.
}
\details{
The input \code{data} has to be a dataframe (or a tibble) with at least the following columns:

\describe{
  \item{client}{\code{character}. Clients Names}
  \item{date}{\code{Date}. Date of the order}
  \item{order}{\code{character}. Code of the order}
  \item{start}{\code{Date}. Start date of the campaign}
  \item{end}{\code{Date}. End date of the campaign}
  \item{season}{\code{character}. Code of the season}
  \item{gender_3bs}{\code{character}. Details for man or women}
  \item{category}{\code{character}. Details of the different types of category}
  \item{style_code}{\code{character}. Code of the style}
  \item{fabric_code}{\code{character}. Code of the fabric}
  \item{color_code}{\code{character}. Code of the color}
  \item{qty}{\code{numeric}. Quantity of the order}
}
}
