% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixcovid19.r
\name{create_dynamic_weight}
\alias{create_dynamic_weight}
\title{Create Dynamic Weights}
\usage{
create_dynamic_weight(
  data,
  season_target,
  seasons,
  current_dates,
  n_week_euclidean,
  weight
)
}
\arguments{
\item{data}{Dataframe. See \code{details}}

\item{season_target}{Character. Target Season.}

\item{seasons}{Character Vector. A vector with the previous Seasons.}

\item{current_dates}{Numeric Vector. Dates in which you want to launch the simulation. Preferable more than one.}

\item{n_week_euclidean}{Numeric. The number of observations we want to calculate from the last one of the season_target season.}
}
\value{
The output is a named numeric vector. Weights to apply for each season when calculating the probs.
}
\description{
This function create dynamic weights for the predictive model and it works as follow:

1) It takes the SKU orders lagged by the function \code{\link{traslate_curves_covid19}}

2) It use \code{\link{slice_data}} and \code{\link{prepare_data_weekly}} in order to obtain the weekly dataframe sliced

3) It takes the value of \code{n_week_euclidean} that is set as default at 5. Calculate the last 5 weeks Euclidean distance
between season_target season and others

4) It uses the inverse-square law to calculate the intensity and then to allocate the weights
}
\details{
The input \code{data} has to be a dataframe (or a tibble) with at least the following columns:

\describe{
  \item{date}{\code{Date}. Date of the order}
  \item{order}{\code{character}. Code of the order}
  \item{end}{\code{Date}. End date of the campaign}
  \item{season}{\code{character}. Code of the season}
  \item{qty}{\code{numeric}. Quantity of the order}
}
}
