% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixcovid19.r
\name{dtw_align_curves}
\alias{dtw_align_curves}
\title{Alignment Curves by Dynamic-Time-Warping}
\usage{
dtw_align_curves(data, season_target, seasons, threshold)
}
\arguments{
\item{data}{Dataframe. See \code{details}}

\item{season_target}{Character. Target Season.}

\item{seasons}{Character Vector. A vector with the previous Seasons.}

\item{threshold}{Numeric. How high want the accurate alignment. Less value more precision.}
}
\value{
Returns a new dataframe with the optimal week lag and is composed by:

\describe{
  \item{season}{\code{character} class. A tibble with the season detail}
  \item{week_lag}{\code{numeric} class. A tibble with the week lag}
}
}
\description{
This function is the core of the project. By an allignment if the curves by dtw package.
This is an internal function of  \code{\link{calculate_lag_data}}. It is not supposed to be called directly
}
\details{
The input \code{data} has to be a dataframe (or a tibble) with at least the following columns:

\describe{
  \item{season}{\code{character} class. A tibble with the season detail}
  \item{week_count}{\code{numeric} class. A tibble with the week difference from the end date}
  \item{qty}{\code{numeric} class. A tibble with the qty by season-date}
  \item{qty_cumsum}{\code{numeric} class. A tibble with the cumulative qty by season-date.}
}
}
