% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixcovid19.r
\name{plot_current}
\alias{plot_current}
\title{Plot Imagines}
\usage{
plot_current(data, current_dates, target, flag_allineamento)
}
\arguments{
\item{data}{Dataframe. See \code{details}}

\item{current_dates}{Numeric Vector. Dates in which you want to launch the simulation. Preferable more than one.}

\item{flag_allineamento}{Boolean. This flag decide if we show alignment or no}

\item{season_target}{Character. Target Season.}
}
\value{
The output is a plot.
}
\description{
This function take all the season and plot them.
}
\details{
The input \code{data} has to be a dataframe (or a tibble) with at least the following columns:

\describe{
  \item{date}{\code{Date}. Date of the order}
  \item{order}{\code{character}. Code of the order}
  \item{end}{\code{Date}. End date of the campaign}
  \item{season}{\code{character}. Code of the season}
  \item{qty}{\code{numeric}. Quantity of the order}
}
}
