% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixcovid19.r
\name{plot_weight}
\alias{plot_weight}
\title{Plot Table Weights}
\usage{
plot_weight(weight)
}
\arguments{
\item{weight}{Vector. This is the vector of weights,}
}
\value{
The output is a plot.
}
\description{
This function take all the season and plot them.
}
