% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixcovid19.r
\name{prepare_data_weekly}
\alias{prepare_data_weekly}
\title{Prepare Dataframe}
\usage{
prepare_data_weekly(data, season_target, seasons)
}
\arguments{
\item{data}{Dataframe. See \code{details}}

\item{season_target}{Character. Target Season.}

\item{seasons}{Character Vector. A vector with the previous Seasons.}
}
\value{
Returns a new dataframe and is composed by:

\describe{
  \item{season}{\code{character} class. A tibble with the season detail}
  \item{week_count}{\code{numeric} class. A tibble with the week difference from the end date}
  \item{qty}{\code{numeric} class. A tibble with the qty by season-date}
  \item{qty_cumsum}{\code{numeric} class. A tibble with the cumulative qty by season-date}

}
}
\description{
This functions transform daily SKU orders by wholesaler in weekly data.
This is an internal function of \code{\link{create_dynamic_weight}}
and \code{\link{calculate_lag_data}}. It is not supposed to be called directly
}
\details{
The input \code{data} has to be a dataframe (or a tibble) with at least the following columns:

\describe{
  \item{client}{\code{character}. Clients Names}
  \item{date}{\code{Date}. Date of the order}
  \item{order}{\code{character}. Code of the order}
  \item{start}{\code{Date}. Start date of the campaign}
  \item{end}{\code{Date}. End date of the campaign}
  \item{season}{\code{character}. Code of the season}
  \item{gender_3bs}{\code{character}. Details for man or women}
  \item{category}{\code{character}. Details of the different types of category}
  \item{style_code}{\code{character}. Code of the style}
  \item{fabric_code}{\code{character}. Code of the fabric}
  \item{color_code}{\code{character}. Code of the color}
  \item{qty}{\code{numeric}. Quantity of the order}
}
}
