% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixcovid19.r
\name{slice_data}
\alias{slice_data}
\title{Slice SKU weekly Dataframe}
\usage{
slice_data(data, current_dates, season_target)
}
\arguments{
\item{data}{Dataframe. See \code{details}}

\item{current_dates}{Numeric Vector. Dates in which you want to launch the simulation. Preferable more than one.}

\item{season_target}{Character. Target Season.}
}
\value{
Returns the dataframe sliced.
}
\description{
This function take SKU modified Dataframe and make a slice of the data.
This is an internal function of \code{\link{create_dynamic_weight}}
and \code{\link{calculate_lag_data}}. It is not supposed to be called directly
}
\details{
The input \code{data} has to be a dataframe (or a tibble) with at least the following columns:

\describe{
  \item{season}{\code{character} class. A tibble with the qty detail by client-season-order-date}
  \item{week_count}{\code{numeric} class. A tibble with the qty detail by season-order-date-sku}
  \item{qty}{\code{numeric} class. A tibble with the cumulative qty by season-date}
  \item{qty_cumsum}{\code{numeric} class. A tibble with the cumulative qty by season-date. These are not used for the simulation.}
}
}
