% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixcovid19.r
\name{traslate_curves_covid19}
\alias{traslate_curves_covid19}
\title{Traslate curves}
\usage{
traslate_curves_covid19(data, data_lag, drop_orders = FALSE)
}
\arguments{
\item{data}{Dataframe. See \code{details}}

\item{data_lag}{Dataframe. See \code{details}}

\item{drop_orders}{Logical. See \code{details}}
}
\value{
Returns SKU orders dataframe lagged.
}
\description{
This functions translate the SKU orders curves with the lag calculate in \code{\link{calculate_lag_data}}.
}
\details{
The input \code{data} has to be a dataframe (or a tibble) with at least the following columns:

\describe{
  \item{client}{\code{character}. Clients Names}
  \item{date}{\code{Date}. Date of the order}
  \item{order}{\code{character}. Code of the order}
  \item{start}{\code{Date}. Start date of the campaign}
  \item{end}{\code{Date}. End date of the campaign}
  \item{season}{\code{character}. Code of the season}
  \item{gender_3bs}{\code{character}. Details for man or women}
  \item{category}{\code{character}. Details of the different types of category}
  \item{style_code}{\code{character}. Code of the style}
  \item{fabric_code}{\code{character}. Code of the fabric}
  \item{color_code}{\code{character}. Code of the color}
  \item{qty}{\code{numeric}. Quantity of the order}
}

The input \code{data_lag} has to be a dataframe (or a tibble) with the following columns:

\describe{
  \item{season}{\code{character} class. A tibble with the season detail}
  \item{week_lag}{\code{numeric} class. A tibble with the week lag}
}

If Drop orders is TRUE then we drop orders that fall between the season end and the estimated lag.
}
